Split VIS mesh into header/definition, added documentation, cleaned up includes
